requirements.txt
flask
gradio
pandas
networkx
sounddevice
scipy
SpeechRecognition
edge-tts
requests
python-dotenv